pyproject.toml
[project]
name = "language-tutor"
version = "0.1.0"
description = "French Practice AI Application with TTS and dialogue generation"
readme = "README.md"
authors = [{ name = "jcf44", email = "johancobo236@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.29.0",
    "openai>=1.3.0",
    "google-generativeai>=0.3.0",
    "google-cloud-texttospeech>=2.16.0",
    "azure-cognitiveservices-speech>=1.34.0",
    "whisperspeech",
    "torch",
    "torchaudio",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydub>=0.25.1",
    "pandas>=2.1.0",
    "requests>=2.31.0",
    "aiofiles>=23.2.0",
    "asyncio-throttle>=1.0.2",
    "webdataset>=0.2.111",
    "gtts>=2.5.4",
]

[project.scripts]
language-tutor = "language_tutor.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
